working on table

+++ app/Http/Controllers/Admin/ProductController.php
@@ -2,10 +2,14 @@
 
 namespace App\Http\Controllers\Admin;
 
+use App\Events\ProductDeleteResponseInAdmin;
 use App\Http\Controllers\Controller;
 use App\Models\Product;
+use App\Models\User;
 use Illuminate\Http\Request;
 use Inertia\Inertia;
+
+use Illuminate\Support\Facades\Auth;
 
 class ProductController extends Controller
 {
@@ -13,6 +17,7 @@ class ProductController extends Controller
     {
         return Inertia::render('products/Index', [
             'products' => Product::all(),
+            'user' => Auth::user(),
         ]);
     }
 
@@ -23,6 +28,25 @@ class ProductController extends Controller
 
     public function destroy(Product $product)
     {
+        
+        // $admin = User::where('is_admin', 1)->first();
+
+        $admin = User::where('is_admin', 1)->first();
+
+        // Check if $admin is found
+        if (!$admin) {
+            \Log::error("No admin user found.");
+            return redirect()->back()->with('error', 'No admin found to handle the broadcast.');
+        }
+        
+        try {
+            \Log::info("Dispatching ProductDeleteResponseInAdmin event for product: {$product->id}");
+            broadcast(new ProductDeleteResponseInAdmin($product, $admin))->toOthers();
+        } catch (\Exception $e) {
+            \Log::error('Broadcast error: ' . $e->getMessage());
+        }
+
+
         $product->delete();
 
         return redirect()->back()->with('success', 'Product Deleted Successfully!');
@@ -40,6 +64,14 @@ class ProductController extends Controller
 
         Product::create($attributes);
 
+
         return redirect()->back()->with('success', 'Product Add Successfully!');
+    }
+
+    public function status(Product $product)
+    {
+        $product->status = $product->status == 'active' ? 'inactive' : 'active';
+        $product->save();
+        return redirect()->back()->with("success", 'Product Status Update Successfully');
     }
 }
